refactor(registry): S3 fixup — move signAdminRequest to bus/stack-provisioning.ts to break the import cycle
Sage review on PR #1532 flagged the first draft's siting of signAdminRequest
in common/registry/signing.ts: that file's canonicalJSON already flows into
bus/stack-provisioning.ts, so importing signClaimWithSeed back the other way
formed a real common<->bus cycle, papered over with a "verified safe by tsc"
comment rather than actually broken.

Moves signAdminRequest into stack-provisioning.ts, next to signClaimWithSeed
(called directly, no import needed), reusing that module's existing
one-directional canonicalJSON import. signing.ts is back to its original
shape apart from a header-comment pointer to where the signer now lives.
Updates the 8 call sites' imports accordingly (bus/stack-provisioning instead
of common/registry/signing).

Moves the byte-identical + live-registry round-trip tests from
common/registry/__tests__/signing.test.ts into
src/bus/__tests__/stack-provisioning.test.ts. Verified via a throwaway
runtime smoke check that the move does not perturb the signed bytes: the
pinned seed+claim fixture produces the exact same signature and wire string
before and after.
